TouchPad now extends Session, move stuff from ws server back to Session
